+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - **MinusDMIndicator** moved to adx-package
 - `analysis/criterion`-package moved to root
 - `cost`-package moved to `analysis/cost`-package
+- **AroonXXX** indicators moved to aroon package
 
 ### Fixed
 - **LosingPositionsRatioCriterion** correct betterThan
@@ -33,6 +34,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 
 ### Added
 - **NumericIndicator** new class
+- **AroonFacade**, **BoliingerBandFacade**, **KeltnerChannelFacade** new classes
 - **AbstractEMAIndicator** added getBarCount() to support future enhancements 
 - **ATRIndicator** "uncached" by changing superclass to AbstractIndicator; added constructor to accept TRIndicator and getter for same; added toString(); added getBarCount() to support future enhancements
 - :tada: **Enhancement** added possibility to use CostModels when backtesting with the BacktestExecutor
