+++ ta4j-core/src/test/java/org/ta4j/core/indicators/AroonUpIndicatorTest.java
@@ -33,6 +33,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.ta4j.core.BarSeries;
 import org.ta4j.core.BaseBarSeries;
+import org.ta4j.core.indicators.aroon.AroonUpIndicator;
 import org.ta4j.core.num.Num;
 
 public class AroonUpIndicatorTest {
